Move AgentInterface to Client::UI::Agent
